Fix pre-commit hook to not require a dependency on env 8.30
Previously, the pre-commit script would give an error on Ubuntu 18.04
and any other OS which did not have an extremely recent version of env:

```
/usr/bin/env: invalid option -- 'S'
Try '/usr/bin/env --help' for more information.
```

The -S parameter was introduced to allow passing arguments to tcsh while
still allowing it to be anywhere in $PATH. However the only argument
being passed is `-f`, which only speeds up the start time slightly and
does not affect the behavior. So it can be removed without breaking
anything.