[#527] CREATE TABLE now supports user-defined SQL NULL characters
Previously, Octo only supported SQL `NULL` values by implicitly treating empty strings as `NULL` based on type and `NOT NULL` specifications in the DDL. This commit enhances Octo to allow users to define specific characters to be treated as `NULL` values using a new `NULLCHAR` keyword in an Octo `CREATE TABLE` statement.

The new syntax takes the form of a `NULLCHAR` keyword, followed by a series of one or more integer ASCII values between 0 and 127 enclosed in parentheses, e.g. `NULLCHAR (127)`. The documentation in `grammar.rst` has been updated to detail the form of this new syntax.

To implement this new feature, both the parser and lexer were updated to recognize the new `NULLCHAR` keyword, with the parser specifically being updated with a new rule for processing this keyword and its arguments.

Since the `NULLCHAR` specification is effective at the table/DDL level, the `SqlTable` struct definition was updated to store this specification in a new `nullchar` member for access during physical planning. The relevant physical plan updates are in `tmpl_column_reference.ctemplate` and `tmpl_tablejoin_body.ctemplate` and build on the `empty2null` M helper function introduced in the resolution to #320.

Specifically, `empty2null` was enhanced to accept a new `nullchar` argument that contains, if specified, the integer value(s) from a `NULLCHAR` specification in the DDL/table definition. Similarly, this function was enhanced with logic for converting any specified `NULLCHAR`s to `$ZYSQLNULL` when encountered.

For reporting errors in `NULLCHAR` specifications encountered in a DDL, the `ERR_INVALID_DELIM_CHAR` error message was revised into a more generic `ERR_INVALID_KEYWORD_CHAR` error message.

`create_table_defaults.c` was also updated to account for the case where `NULLCHAR` is not specified in the DDL. Similarly, `compress_statement.c` and `decompress_statement.c` were enhanced to include the new `nullchar` field of `SqlTable` structs when creating binary table definitions.

These changes are tested by new test cases added to `test_createtable.bats.in`, which use the Octo-Postgres crosscheck method to validate results. To support these new tests, new fixtures were added for Octo and Postgres: `nullcharnames.sql`, `nullcharnames.zwr`, and `postgres-nullcharnames.sql`. The latter is now included in `hello_psql.bats.in` as well.

Additionally many outrefs were updated to account for the physical plan revisions mentioned above as well as the new `postgres-nullcharnames.sql` fixture (`HP01.ref`):
+ `HP01.ref`
+ `TC027.ref`
+ `TC028.ref`
+ `TCJ11.ref`
+ `TCK03.ref`
+ `TCT011.ref`
+ `TEO03.ref`
+ `TGB04.ref`
+ `TJC003.ref`
+ `TML010.ref`
+ `TML011.ref`
+ `TMU02.ref`
+ `TWI01.ref`
+ `TWI02.ref`
+ `TWI03.ref`
+ `TWI04.ref`
+ `TWI05.ref`
+ `TWI06.ref`
+ `TWI07.ref`
+ `TWI09.ref`
+ `TWI12.ref`
+ `TWO01.ref`
+ `TWO02.ref`
+ `TWO03.ref`
+ `TWO04.ref`
+ `TWO05.ref`
+ `TWO06.ref`
+ `TWO07.ref`
+ `TWO08.ref`
+ `TWO09.ref`
+ `TWO10.ref`
+ `TWO11.ref`
+ `TWO12.ref`
+ `TWO13.ref`
+ `TWO14.ref`
+ `TWO15.ref`

Miscellaneous change:
+ Removed obsoleted parser rule for boolean comparisons