[#539] Require psql queries to succeed
Previously errors would be silently ignored.

- Don't use ON_ERROR_STOP on CentOS 7

On CentOS 7, some NULLs are not correctly inferred to the wrong type.
Rather than modifying many dozens of tests, this allows queries to fail
on machines with too low a version of Postgres.

- Use postgres database instead of the default for detecting the database version

In pipeline runs, there is no database set up for the root user.