[#531] Don't remove NULL checks for outer joins
Consider the following query:

```sql
SELECT c.customer_id FROM orders o RIGHT JOIN customers c ON '1' = c.zipcode
WHERE (o.order_id IS NOT NULL);
```

The correct output is 0 rows: because table o is on the left side of the
RIGHT join and because no customer has a zipcode of 1, all columns of o
that get chosen due to the RIGHT join will have a value of NULL. Since
the WHERE clause has `(o.order_id IS NOT NULL)` it should filter out all
those chosen rows resulting in 0 rows.

Instead Octo would previously give all 5 customer ids:

```sql
OCTO> select c.customer_id from orders o right join customers c on '1' = c.zipcode where (o.order_id IS NOT NULL);
1
2
3
4
5
OCTO>
```

The issue is that although order_id is a primary key for orders, it is
not a primary key for the right join. The optimization can only occur if
there are no LEFT, RIGHT, or FULL JOINs (if there is an INNER JOIN, then
there will still be no case when the primary key is not null).

- Add the query which originally failed in the pipeline
- Improve the comments for the optimization
- Add description of the test to the reference file