[#533] Updated get_ydb_release.sh to avoid ZROSYNTAX and permissions errors when attempting to use _XCMD routine
Previously, running `get_ydb_release.sh`, i.e. implicitly when building with CMake, with `$ydb_routines` pointing to a yet ungenerated Octo routine file resulted in `ZROSYNTAX` errors from CMake. That was initially corrected by setting `ydb_routines=$ydb_dist` in `get_ydb_release.sh`. However, this does not work in the case where the `_XCMD` command is to be found in `libyottadbutil.so`.

Accordingly, this commit fixes the issue by adding a check in `get_ydb_release.sh` for the existence of `libyottadbutil.so` and sets `ydb_routines=$ydb_dist/libyottadbutil.so` in that case.