Revised Octo installation testing procedures to prevent failures
Previously, Octo installation would entail the execution of a subset of the full test suite, including `test_select_columns.bats.in`, which fails when `psql` is not installed due to `TSC20`. This failure is now avoided when `psql` is unavailable during installatoin by moving `TSC20` to `TSCP06` so that it runs with related tests requiring `psql`.

Additionally, the Octo installation tests are now optional and disabled by default. They can be enabled when running `cmake` by setting `ENABLE_INSTALL_TESTS=ON`.