[#420] Corrected run_multiple_query_files.java to run specified protocol
Previously (per fixes for #420), `run_multiple_query_files.java` used the `==` operator to compare a string literal with a `String` object, leading to the usage of the Simple Query protocol even when the Extended Query protocol was specified. This did not cause test failures, but silently reduced test coverage.

This is now fixed by replacing this comparison with a call to the preferred `String.equals` method for `String` objects.