[#527][#320] Include empty2null conversion when DDL specifies $EXTRACT
Previously, plan generation did not account for the possible inclusion of an `EXTRACT` specification in an Octo DDL when handling empty string to SQL `NULL` conversions, i.e. the `empty2null` helper routine was not called in this case. This led to fields not being converted to `NULL` when they should be. This commit corrects this behavior by updating `tmpl_column_reference.ctemplate` to apply `empty2null` even when an `EXTRACT` is present in the DDL.

A new `TC032` test was added to `test_createtable.bats.in` to validate this behavior, along with new `TC032` fixtures. This test uses the JDBC driver to validate results, as this client will automatically report SQL `NULL` fields as `null`. A new `TC034` test was added as well, to verify correct behavior when `NULLCHAR` is specified in the DDL.

Note that the above test revealed an issue in rocto's handling of `CREATE TABLE` statements within the Extended Query protocol. Specifically, the `sql_schema_statement` parser rule did not account for possible Extended Query protocol usage and so did not set the `parse_context->routine` variable to `"none"` as is done for other non-`SELECT` statements such as `SET` and `SHOW`. This led to an attempt in `handle_bind.c` to populate bind parameters for a plan that doesn't exist (no M plan is generated for `CREATE TABLE` statements) and in turn to an assert failure.

To correct this, the `sql_schema_statement` parser rule now sets `parse_context->routine` to `"none"`, indicating to the subsequent Extended Query logic to correctly omit parameter population and plan processing for `CREATE TABLE` statements.