[#435] Added support for identifiers in SET statements
Previously, Octo only accepted literals in SQL `SET` statements, but not identifiers, leading to incompatibilities with some clients such as PGAdmin4. Now, Octo accepts both identifiers and literals in `SET` statements. This was achieved by adding two new rules to `set.y` to handle the two possible `SET` syntax cases (`TO` and `=`). To reduce duplication in these rules, a new `parser/sql_set_statement.c` module was added.

Additionally, a bug was discovered in `print_temporary_table.c` that led to the results of database lookups for `SHOW` statements to omit the null terminator. This is now fixed.

A new `TJC013` test case was added to validate the new behavior.