[#483] Added column-level delimiter support to Octo DDL
Previously, Octo would accept table-level delimiter specifications, but not column-level ones. This is now fixed so that Octo accepts both types of delimiter specifications.

This was done by adding a new `delim` field to the `SqlColumn` struct and updating this field in the parser when it is specified, or else setting it to `NULL`. To facilitate this and reduce duplication, previous `DELIM`-related rules were combined into a single `delim_specification` rule that is now used in both the `optional_keyword_element` (table-level) and `column_definition_tail` (column-level) rules.

To make use of this new column-level delimiter field, `tmpl_column_reference.ctemplate` was updated to retrieve it and apply the specified delimiter when present, overriding the table-level delimiter for just the given column. Also, `compress_statement.c` and `decompress_statement.c` were updated to store and retrieve this field, respectively, during DDL creation and loading.

Finally, a new `TC035` test case was added to validate the new behavior.