[#592] Added mnemonic error name to error output and error messages to docs
Previously, Octo error messages omitted the mnemonic error name used to identify the error internally and in the documentation, though it would be more user-friendly to include the former rather than omit it. Accordingly, Octo now includes the mnemonic error name in each error message.

This was accomplished by revising `errors.c` to include the mnemonic error name in the error message prefix strings specified for Octo and Rocto there. Similarly, a new `err_name_str` array is now automatically generated in this file from the error definitions in `errors.hd`, similar to how the error message format strings themselves are generated.

Additionally, a new `INFO_ENV_VAR` error message was added in replacement of the `CUSTOM_ERROR` used in `octo_init.c` for outputting environment variable information.

Also, as requested in #592, error format messages were added to `errors.rst`. To make it possible to update this by a single script and to prevent confusion with `ERR_SEND_MESSAGE`, `ERR_ROCTO_SEND_MESSAGE` was renamed to `ERR_ROCTO_SEND_FAILED`. The usages of this message and its documentation were updated accordingly.

To accommodate the new error message format, a large number of outrefs were updated to include the mnemonic name for each recorded error message.

Some other test updates were also required, namely:
+ Revised `verify_output` helper function to include `INFO_ENV_VAR` when stripping environment variable output
+ Revised `test_psql_authentication.bats` to use username other than 'jon' to prevent conflicts with developer environment
+ Corrected test number in `test_index_statement.bats.in` to `TIS001`
+ Revised `cut` command in `TMU02` to account for additional spaces in output due to mnemonic inclusion
+ Revised `grep` calls in several test cases in `test_squirrel_sql_connect_queries.bats.in` to account for mnemonic error names

Additionally, two new scripts were added to verify that all errors are accounted for in the documentation and to automatically add error message text if absent. The first, `doc_error_update.sh` performs both of these functions and is intended to be run as part of the `pre-commit` hook, which was also updated. The latter, `doc_error_check.sh` only performs the verification step and runs as part of the pipeline in `build.sh`, which was modified to include it. Note also that a new `omitted_errors.ref` file was added for use in both scripts for tracking error message special cases and omitting them from these scripts due to exceptional error mnemonics or message text.

The addition of the above scripts revealed that a number of error messages were completely omitted from the documentation. These missing messages were added as part of this commit. Similarly, a redundant `ERR_ROCTO_SEND_MESSAGE` was combined with `ERR_ROCTO_SEND_FAILED` and removed.

Also, some error messages were revised to use single quotes instead of escaped double quotes for the sake of more uniform error messages and consistency with the above scripts.

A number of other related changes were made:
+ Added `ERR_` and `INFO_` prefixes to documentation error mnemonics
+ Added `Description/Action` prefix to error message descriptions
+ Revised `ERR_CURPLAN` and `ERR_CLIENT_CONNECTED` to `INFO_` prefix
+ Replaced various `CUSTOM_ERROR` calls with named error messages, updated `errors.hd` accordingly
+ Converted several trace category messages to `INFO_` prefix from `ERR_`
+ Added `ROCTO_` prefix to errors.rst to match actual error code mnemonic