[#90] Added new ERR_DDL_LITERAL message for clarifying DDL syntax errors
Previously, the `ddl_int_literal_value` and `ddl_string_literal_value` parser rules emitted a generic syntax error when invalid types are used in certain DDL keyword specifications. However, a more specific message would be more user-friendly and so is added with this commit in the form a new `ERR_DDL_LITERAL` message.

A few outrefs were updated to account for this new message. Similarly, this message was added to the documentation.

Additionally, a missing `assert` was added to an error case in the `ddl_string_literal_value` parser rule, as all `$LITERAL`s should result in `value_STATEMENT`s and so that error case should never occur.