[#592] Fixed `doc_error_update.sh` issues
Previously, the `doc_error_updated.sh` script did not work correctly when included the `pre-commit` hook, resulting in duplicate error messages and erroneously modified files. These issues are corrected by this commit.

Specifically, the determination of the format string text was corrected to omit the double-quotes from the format string when it is extracted from `errors.hd`. Including these quotes causes messages to not be detected when they are present, as `errors.rst` does not include quotes in its message text.

Also, a list is now compiled that includes each format text string missing from `errors.rst` and the script fails if this list non-empty. Nonetheless, the previous behavior of updating `errors.rst` with these missing strings is retained. The result is that, these changes are applied, but the commit fails when format text is missing for an error message. This allows the commit author to review the automatically applied changes before committing.

Finally, a new check for duplicated error messages was added to `doc_error_update.sh`. To pass this new check, a couple revisions were made:

+ Removed duplicate `INFO_MEMORY_USAGE` from `errors.rst`
+ Revised `ERR_ROCTO_INVALID_CHAR_VALUE to `ERR_ROCTO_INVALID_ITEM_VALUE`
+ Revised `ERR_ROCTO_INVALID_INT_MULTI` to `ERR_ROCTO_INVALID_FORMAT_CODE`