Added note using CTRL+C to exit octo/rocto to docs
It was discovered that `rocto` listener instances will not exit in response to `CTRL+C` if there are active client network connections. This is expected behavior and so there is no bug here. However, it would be helpful to document this behavior and so that is done in this commit.