Only run the pipeline if the doc/ folder is modified
This has the caveat that GitLab is not very good at telling what
'modified' means. From [the docs]:

> Without pipelines for merge requests, pipelines run on branches or tags that don’t have an explicit association with a merge request. In this case, a previous SHA is used to calculate the diff, which is equivalent to git diff HEAD~.

In particular, this means that if there are multiple commits in an MR,
the latest commit does not modify `doc/`, but an earlier commit does,
then the YDBDoc pipeline will not run when the MR is merged.

Since most MRs are only one commit and it's still possible to trigger a
manual rebuild of YDBDoc, this seems like a reasonable tradeoff to me.
The alternative is to always trigger a YDBDoc rebuild, which rebuilds
more often than necessary, but never misses a rebuild.

Note that the alternative 'Using only:changes with pipelines for merge
requests' doesn't work because YDBDoc always downloads the master branch
of YDBOcto, but `only: merge_requests` runs before the MR is merged into
master.

[the docs]: https://docs.gitlab.com/ee/ci/yaml/#using-onlychanges-without-pipelines-for-merge-requests