[#565] Removed rounding of `precision` variable in `ROUND` routine of `_ydboctosqlfunctions.m` and comment referencing #565
Per a comment in the in `ROUND` routine of `_ydboctosqlfunctions.m`, the logic for this routine could be simplified by removing the rounding of `NUMERIC` arguments once function argument types are properly enforced (#565). Since #565 is resolved, this simplification was done with this commit.

As part of this change, `octo-seed.sql` was updated to no longer accept `NUMERIC` precision arguments in the `ROUND` SQL function, as it should only accept precisions specified as integers.

To account for the two removed `CREATE FUNCTION` statements described above, a number of outrefs were updated to reflect the omission of these functions from the PostgreSQL catalog. Also, `TSCP04_errors.ref` to reflect the `UNKNOWN_FUNCTION` error messages that are now issued when users attempt to call one of the removed `ROUND` interfaces.