[#637] Octo does not process interrupts until it finishes parsing a statement
As part of YottaDB/DB/YDB#560, Ctrl-C (`SIGINT`) handling was changed so the process does not terminate
immediately but instead waits for a logical point at which time it terminates.

The application using YottaDB (Octo in this case) needs to call `ydb_eintr_handler()` at the logical
point to finish handling the Ctrl-C.

This was not done in YDBOcto as part of YottaDB/DB/YDB#560 and so a `Ctrl-C` pressed at the `OCTO>`
prompt did not terminate the process even though a newline was pressed.

This is now fixed to invoke `ydb_eintr_handler()` right after `readline()` call returns. This means
that a Ctrl-C would terminate the Octo process only after a newline is pressed. This is better than
before where a multi line query had to be terminated with a `;` before the Ctrl-C terminated the process.

One might argue that a Ctrl-C should terminate the process right away instead of waiting for a newline
to be entered. But that does not seem a simple change as we are inside libreadline() code at that point
and we need to define signal handlers in Octo for libreadline to invoke. Not sure if this effort is worth it
at this point.