[#636] Fix improper check for TC011 subtest in tools/ci/build.sh (introduced in previous commit)
As part of a prior commit (5b08cc0b9fd5f61f2ed733460e470db8f9550f8a), `tools/ci/build.sh` was fixed
to check if the current subtest is `TC011` and if so skip rerunning that in the `test-auto-upgrade`
pipeline job.

But the check was not right. The `subtest` variable would have the following value whereas the check
expected the variable to have just the `TC011` value.

```
TC011 : OCTO411 : Add support for INTEGER precision specification to parser for CREATE TABLE statements`
```

This is fixed so the check now does a regex check (`=~`) instead of an equality check (`==`).