Fixed pipeline failure to correctly execute `doc_error_update.sh`
Previously, `build.sh` would call `doc_error_update.sh` from the build directory, whereas the latter expects to called from the top-level YDBOcto repository directory for compatibility with the git `pre-commit` hook that runs the same script. This resulted in supporting files located in `tools/ci` to not be found and quietly fail to perform the desired error message documentation check. This was resolved by adding `pushd`/`popd` instructions around the `doc_error_update.sh` call in `build.sh` to ensure the correct directory path is used.

Additionally, it was discovered that `status` may be a `bash` reserve word (as indicated through a syntax checker), and so this was renamed to `result` in `doc_error_update.sh` to prevent the possibility of quiet failures related to this variable not behaving as expected.

Also, `doc_error_update.sh` was updated to include `set -e` to force script failure on any failed command as well as a top-level repository directory check to confirm it is being run from the correct directory.

Finally, `ERR_LITERAL_MAX_LEN` was added to `docs/errors.rst` to satisfy the now correctly functioning pipeline check.