[#641] Don't rebuild `octo_conf.c` unnecessarily
Previously, the dependency was stated as `default_octo_conf.h`,
which expanded to `src/default_octo_conf.h`, in the source directory.
However the file was generated in the build directory, so make would
always think it was out of date. This changes the dependency to look at
the build directory instead.

The new output looks like this:

```
$ make
[  0%] Built target libhelpers
[  3%] Built target pparser
[ 70%] Built target libocto
[ 91%] Built target librocto
[ 93%] Built target rocto
[ 98%] Built target _ydbocto
[100%] Built target octo
[100%] Built target convert_catalog_zwr
[100%] Built target convert_catalog_sql
```