[#643] ORDER BY returns incorrect results when used on a value returned by the ROUND function
Background
-----------
Below is a simple example illustrating the issue (it was originally discovered by @zylog1O1 while
fixing !829 when ROUND of a floating point number returned by the AVG function was added to a query
that already had ORDER BY usage).

Below is a query that returns the floating point numbers in the right order.

```sql
OCTO> select id from (values (ROUND(2.1234567890,10)), (ROUND(1,10))) as n1(id) order by id;
1.0000000000
2.1234567890
```

But the below query (which differs from the above in only the 10th digit after the decimal point and
expects the same output as the first query) returns floating point numbers in the wrong order.

```sql
OCTO> select id from (values (ROUND(2.1234567891,10)), (ROUND(1,10))) as n1(id) order by id;
2.1234567891
1.0000000000
```

This is because the `ROUND` function currently returns a numeric quantity formatted as a string. And
when used with an ORDER BY, we should take its numeric value before trying to order it. Taking the
string value as is will cause problems due to canonical number issues with the underlying YottaDB
database. See below example that demonstrates this counter-intuitive behavior of YottaDB with numbers
stored as strings.

When numerics are specified as is in subscripts, they show up in the right order.

```m
YDB>kill x
YDB>set x(10282.6666666667)=1
YDB>set x(10273.0000000000)=1
YDB>zwrite x
x(10273)=1
x(10282.6666666667)=1
```

But when numerics are specified as strings in subscripts, they show up out of order.

```m
YDB>kill x
YDB>set x("10282.6666666667")=1
YDB>set x("10273.0000000000")=1
YDB>zwrite x
x(10282.6666666667)=1
x("10273.0000000000")=1
```

It is not easily possible to fix YottaDB's behavior so the fix has to be in Octo. When an ORDER BY
is done on a value that is a numeric type (such as those returned by the `ROUND` function), it should
generate code to automatically convert the underlying value to a numeric (using the `+` operator).

Fix
----
* The fix is in `src/m_templates/tmpl_column_list_combine.ctemplate` where if we are generating M code
  for the `ORDER BY`, we make sure any numeric type quantity that is being ordered is forced into
  a numeric type in M using the `+` operator.

* A new `test_order_by/TOB14` bats subtest verifies this fix.

* A few existing reference files were modified to reflect the slightly changed generated M plans.