[YottaDB/Util/YDBPosix#5] Update pipeline to remove commands for separate UTF-8 install of YDBPosix
This commit updates the pipeline to only install YDBPosix once to reflect the changes to YDBPosix that no longer require a separate install for the UTF-8 version of the plugin. The CMakeLists for YDBPosix now installs both the M mode and UTF-8 mode versions with a single command, if YottaDB is installed with UTF-8 support.