Improve documentation and behavior of clang-tidy in the pipeline
Narayanan recently found that clang-tidy was generating spurious
warnings in the pipeline that looked like this:

```
warning: unknown warning option '-Wmisleading-indentation'; did you mean '-Wbinding-in-condition'? [clang-diagnostic-unknown-warning-option]
```

Despite that, the pipeline was not failing. This fixes the warnings,
documents why the pipeline did not fail, and gives instructions for
replicating the behavior locally.