[#659] Fix bug where Octo would strip trailing zeros from the YDB version
Previously:

```
YottaDB version: r1.3
```

Fixed:

```
YottaDB version: r1.30
```

This uncovered another bug: constants.h will not be automatically
regenerated when `get_ydb_release.sh` is changed. Since
`YDB_RELEASE_VERSION` is set at configure time, not build time, I didn't
see a simple way to fix this, so I judged that this happens infrequently
enough not to be worth fixing.

There was yet another bug: `tools/get_ydb_release.sh` should have caught
that 1.3 < 1.30 and given an error. It was comparing the versions as
floats, not as semantic versioning; the fix was to split on . and compare
each component individually.