[#655] python2 to python3 conversion for test fixtures
Python 2 is being used in tests but is now deprecated.

filter.py and split_queries.py only needed to be switched to python3 on
the first /usr/bin/env line. No code changes were needed as all code was
already legal python3 code.

For insert.py, additional changes were needed to be done to convert maps
to list comprehensions.

The code changes were done by the 2to3 tool
(https://docs.python.org/3/library/2to3.html) and then manually
reviewed.

The following tests used the python code and were re-ran to ascertain
that the no regressions were introduced:

split_queries.py: test_createtable (7 failures on master without python2)
filter.py and insert.py: test_sqllogic (5 failures on master without python2)

Misc Changes
------------

Tangentially related to the subject matter of this commit, Dockerfile-Tests
needed to be updated to to use python3 instead of python2. However,
since it was initially written, a lot more tests have been added and it
has not been updated since then. It took me a lot of time, but I was
able to fix all the problems that it had, which were:

- clang-tidy needed to be installed.
- Debian no longer includes certificates. ca-certificates needed to be
installed. wgetting go didn't work (this should change in any case in
the future to install from debian).
- netstat was not available. net-tools needed to be installed.
- Java as not installed. openjdk-8-jdk-headless needed to be installed.
- expect needed to be installed.

The tests now all pass, except for these:

```
45/123 Test: test_createtable
 not ok 24 TC022 : OCTO320 : Octo converts empty column values based on type when NOT NULL is specified in the DDL
 not ok 25 TC023 : OCTO320 : Octo treats empty column values as NULL when NOT NULL is not specified in the DDL
 not ok 26 TC024 : OCTO320 : Rocto converts empty column values based on type when NOT NULL is specified in the DDL
 not ok 27 TC025 : OCTO320 : Rocto treats empty column values as NULL when NOT NULL is not specified in the DDL
 not ok 29 TC027 : OCTO527 : CREATE TABLE accepts user-defined NULL character and enforces correct precedence when used with NOT NULL
 not ok 37 TC036 : OCTO90 : Simple tests of auto upgrade of binary table/function/plan definitions/artifacts
```

Multiple runs of these shows that they are probably related to postgres
not being configured correctly.

.gitlab-ci.yml had a cludge to make the python3 callable as python. That
is incorrect since the syntax between the python versions can be
slightly different, as this commit demonstrates. Instead, we added
python2 to become the proper python if calling the older python is
needed.

Note: Dockerfile-Tests.centos will probably need updates, and is not
addressed in this commit.