[#655] Only install python2 for test-auto-upgrade
We use python3 in our tests, but previous commits require python2, and
test-auto-upgrade uses previous commits.

The previous commit for the python issue got around this by installing
python2 for every CI that uses ubuntu. @nars1 requested this to be
changed to only apply to the test that uses it.