[#575] Octo CREATE TABLE naively creates invalid GVNs from table names
This commit changes the table name default global storage (if not
defined using the GLOBAL keyword) to store it in %ydboctoD + $ZYSUFFIX
hash (22 long) of the name of the table. This way we are assured of
valid global names.

Previously, table names that contained "_" and "." were allowed as the
global names.

An unrelated change is to reduce code duplication by moving the snprintf
operation and buffer resizing into a new macro: SNPRINTF_TO_BUFFER.
There was a macro called UNUSED to prevent compiler warnings -- that is
no longer needed as once the code is moved into the macro, no warnings
are issue.

Code Changes
------------
`src/create_table_defaults.c`: MMR hash table name and convert to string
instead of just using the table name directly.

Doc Changes
-----------
`doc/grammar.rst`: Section on using GLOBAL updated to describe the new
default and to clarify how to use GLOBAL with CREATE TABLE.

Test Changes
------------
TC044 is a new test to verify the functionality this issue fixes. In
addition, due to the large number of failing tests, some tests were
changed to use the new default global that is created for the table, and
others were changed to explicitly specify the global. See below for
which is which.

* `tests/test_createtable.bats.in`:
	* TC002 - ^NAMES4 ZWR file moved to ^($ZYSUFFIX("NAMES4"))
	* TC004 - Specify T1 global explicitly (`TC004.sql`)
	* TC008 - Specify T1 global explicitly
	* TC009 - Specify T1 global explicitly
	* TC010 - Specify ^DATETIME storage global using $ZYSUFFIX
	* TC012 - Specify SIMPLE[0-2] storage global using $ZYSUFFIX
	* TC033 - Update ref file to reflect that storage is in hashed global
	* TC044 - New test to verify that . or _ in table name doesn't
	  cause an error.
* `tests/test_drop_table.bats.in`
	* TDT02 - Specify T1 global explicitly
* `tests/test_hash_query.bats.in`
	* THQ03 - Specify T1 global explicitly
* `tests/test_exists_operator.bats.in`
	* TEO03
		* `tests/fixtures/sqllogic1.sql` points to T1 global explicitly
		* ref file contains the constructed M code for the
		queries concatenated together using `cat _ydb*.m` . Due to the previous
		change, the routine name is different, so the order of
		the cat output is now different. The actual contents of
		the files were verified to be the same. That took a lot of
		time to troubleshoot!
* `tests/test_large_queries.bats.in`
	* Setup: `largecolnames.sql` global specified explicitly
* `tests/test_primary_key.bats.in`
	* TPK01 - Specify T1 global explicitly
* `tests/test_psql_connection.bats.in`
	* TPC002 - Specify T1 global explicitly, update ref file
	* TPC003 - Specify T1 global explicitly, update ref file